Split Catapult types into separate file
