+++ hls4ml/backends/fpga/fpga_types.py
@@ -248,24 +248,12 @@ class QuartusArrayVariableDefinition(VariableDefinition):
         )
 
 
-class CatapultArrayVariableDefinition(VariableDefinition):
-    def definition_cpp(self, name_suffix='', as_reference=False):
-        return '{type} {name}{suffix}[{shape}] /* {pragma} */'.format(
-            type=self.type.name, name=self.name, suffix=name_suffix, shape=self.size_cpp(), pragma=self.pragma
-        )
-
-
 class VivadoInplaceArrayVariableDefinition(VariableDefinition):
     def definition_cpp(self):
         return f'auto& {self.name} = {self.input_var.name}'
 
 
 class QuartusInplaceArrayVariableDefinition(VariableDefinition):
-    def definition_cpp(self):
-        return f'auto& {self.name} = {self.input_var.name}'
-
-
-class CatapultInplaceArrayVariableDefinition(VariableDefinition):
     def definition_cpp(self):
         return f'auto& {self.name} = {self.input_var.name}'
 
@@ -297,11 +285,6 @@ class QuartusArrayVariableConverter(ArrayVariableConverter):
         super().__init__(type_converter=type_converter, prefix='Quartus', definition_cls=QuartusArrayVariableDefinition)
 
 
-class CatapultArrayVariableConverter(ArrayVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(type_converter=type_converter, prefix='Catapult', definition_cls=CatapultArrayVariableDefinition)
-
-
 class VivadoInplaceArrayVariableConverter(ArrayVariableConverter):
     def __init__(self, type_converter):
         super().__init__(type_converter=type_converter, prefix='Vivado', definition_cls=VivadoInplaceArrayVariableDefinition)
@@ -314,26 +297,12 @@ class QuartusInplaceArrayVariableConverter(ArrayVariableConverter):
         )
 
 
-class CatapultInplaceArrayVariableConverter(ArrayVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(
-            type_converter=type_converter, prefix='Catapult', definition_cls=CatapultInplaceArrayVariableDefinition
-        )
-
-
 # endregion
 
 # region StructMemberVariable
 
 
 class QuartusStructMemberVariableDefinition(VariableDefinition):
-    def definition_cpp(self, name_suffix='', as_reference=False):
-        return '{type} {name}{suffix}[{shape}]'.format(
-            type=self.type.name, name=self.member_name, suffix=name_suffix, shape=self.size_cpp()
-        )
-
-
-class CatapultStructMemberVariableDefinition(VariableDefinition):
     def definition_cpp(self, name_suffix='', as_reference=False):
         return '{type} {name}{suffix}[{shape}]'.format(
             type=self.type.name, name=self.member_name, suffix=name_suffix, shape=self.size_cpp()
@@ -366,13 +335,6 @@ class QuartusStructMemberVariableConverter(StructMemberVariableConverter):
     def __init__(self, type_converter):
         super().__init__(
             type_converter=type_converter, prefix='Quartus', definition_cls=QuartusStructMemberVariableDefinition
-        )
-
-
-class CatapultStructMemberVariableConverter(StructMemberVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(
-            type_converter=type_converter, prefix='Catapult', definition_cls=CatapultStructMemberVariableDefinition
         )
 
 
@@ -409,21 +371,6 @@ class QuartusInplaceStreamVariableDefinition(VariableDefinition):
         return f'auto& {self.name} = {self.input_var.name}'
 
 
-class CatapultStreamVariableDefinition(VariableDefinition):
-    def definition_cpp(self, name_suffix='', as_reference=False):
-        if as_reference:  # Function parameter
-            return f'ac_channel<{self.type.name}> &{self.name}{name_suffix}'
-        else:  # Declaration (string name arg not implemented in ac_channel)
-            return 'ac_channel<{type}> {name}{suffix}/*("{name}")*/'.format(
-                type=self.type.name, name=self.name, suffix=name_suffix
-            )
-
-
-class CatapultInplaceStreamVariableDefinition(VariableDefinition):
-    def definition_cpp(self):
-        return f'auto& {self.name} = {self.input_var.name}'
-
-
 class StreamVariableConverter:
     def __init__(self, type_converter, prefix, definition_cls):
         self.type_converter = type_converter
@@ -453,11 +400,6 @@ class VivadoStreamVariableConverter(StreamVariableConverter):
 class QuartusStreamVariableConverter(StreamVariableConverter):
     def __init__(self, type_converter):
         super().__init__(type_converter=type_converter, prefix='Quartus', definition_cls=QuartusStreamVariableDefinition)
-
-
-class CatapultStreamVariableConverter(StreamVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(type_converter=type_converter, prefix='Catapult', definition_cls=CatapultStreamVariableDefinition)
 
 
 # endregion
@@ -490,13 +432,6 @@ class QuartusInplaceStreamVariableConverter(InplaceStreamVariableConverter):
     def __init__(self, type_converter):
         super().__init__(
             type_converter=type_converter, prefix='Quartus', definition_cls=QuartusInplaceStreamVariableDefinition
-        )
-
-
-class CatapultInplaceStreamVariableConverter(InplaceStreamVariableConverter):
-    def __init__(self, type_converter):
-        super().__init__(
-            type_converter=type_converter, prefix='Catapult', definition_cls=CatapultInplaceStreamVariableDefinition
         )
